requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.3.0
google-genai>=1.0.0
google-cloud-texttospeech==2.16.3
pydub==0.25.1
python-dotenv==1.0.1
